Move large number of functions out of __init__.py files, put them to a separate file to make the init cleaner
