+++ pythainlp/__init__.py
@@ -43,6 +43,7 @@ from pythainlp.tokenize import (
     Tokenizer,
     sent_tokenize,
     subword_tokenize,
+    syllable_tokenize,
     word_tokenize,
 )
 from pythainlp.transliterate import romanize, transliterate
