+++ pythainlp/tokenize/crfcut.py
@@ -21,7 +21,7 @@ import pycrfsuite
 from pythainlp.corpus import corpus_path
 from pythainlp.tokenize import word_tokenize
 
-ENDERS = {
+_ENDERS = {
     # ending honorifics
     "ครับ",
     "ค่ะ",
@@ -74,7 +74,7 @@ ENDERS = {
     "เมื่อไหร่",
     "เมื่อไร",
 }
-STARTERS = {
+_STARTERS = {
     # pronouns
     "ผม",
     "ฉัน",
@@ -147,12 +147,12 @@ def extract_features(
     doc_ender = []
     doc_starter = []
     for i in range(len(doc)):
-        if doc[i] in ENDERS:
+        if doc[i] in _ENDERS:
             doc_ender.append("ender")
         else:
             doc_ender.append("normal")
 
-        if doc[i] in STARTERS:
+        if doc[i] in _STARTERS:
             doc_starter.append("starter")
         else:
             doc_starter.append("normal")
